backend/requirements.txt
flask
flask-cors
openai
langchain
beautifulsoup4
requests
python-dotenv
tiktoken
